@@ -1,25 +1,33 @@
 {
+  "engines": {"node": "8"},
   "name": "functions",
   "description": "Cloud Functions for Firebase",
   "scripts": {
-    "lint": "./node_modules/.bin/eslint .",
+    "lint": "./node_modules/.bin/eslint ./src",
     "serve": "firebase serve --only functions",
     "shell": "firebase experimental:functions:shell",
+    "build": "babel src/index.js -o index.js; babel src/lib -d lib",
     "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
+    "deploy": "npm run build; firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
   "dependencies": {
-    "@google-cloud/functions-emulator": "^1.0.0-alpha.29",
-    "firebase-admin": "^5.8.1",
-    "firebase-functions": "^0.8.1",
-    "google-cloud": "^0.57.0",
-    "moment": "^2.20.1",
-    "moment-range": "^3.1.1"
+    "@google-cloud/storage": "^1.7.0",
+    "child-process-promise": "^2.2.1",
+    "firebase-admin": "6.0.0",
+    "firebase-functions": "^2.0.5",
+    "generate-google-calendar-url": "^0.0.2",
+    "moment": "^2.22.2",
+    "moment-range": "^4.0.1",
+    "nodemailer": "^4.6.8",
+    "sharp": "0.20.5"
   },
   "devDependencies": {
-    "eslint": "^4.12.0",
-    "eslint-plugin-promise": "^3.6.0"
+    "babel-cli": "^6.26.0",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "eslint": "^5.4.0",
+    "eslint-plugin-promise": "^4.0.0"
   },
   "private": true
 }
