@@ -1,89 +1,90 @@
-import React, { Component } from 'react'
-import moment from 'moment'
-import firebase from 'firebase'
-import Month from './Month'
-import DayBig from './DayBig'
-
-const reservationsRef = firebase.database().ref("/reservations")
+import React, {Component} from "react"
+import {Route} from "react-router-dom"
+import moment from "moment"
+import Month from "./Month"
+import DayBig from "./DayBig"
+import {routes, toRoute} from "../../../utils"
+import {RESERVATIONS_FS} from "../../../utils/firebase"
 
 export default class Calendar extends Component {
 
   state = {
-    reservations: null,
-    isDayBig: false,
+    reservations: {},
     date: null,
-    bigDayreservations: null,
-    currentDate: moment()
+    currentDate: moment(),
+    mutat: []
   }
 
 
   componentDidMount() {
-    reservationsRef.on("value", snap => {
-      const reservations = snap.val()
-      reservations && Object.keys(reservations).forEach(reservation => {
-        if(!reservations[reservation].metadata.handled){
-          delete reservations[reservation]
-        } 
+    this.updateURL(this.props.match.params)
+    RESERVATIONS_FS.where("handled", "==", true).onSnapshot(snap => {
+      this.setState({reservations: {}})
+      snap.forEach(reservation => {
+        const {
+          from, to, roomId
+        } = reservation.data()
+        this.setState(({reservations}) => ({reservations: {
+          ...reservations,
+          [reservation.id]: {
+            from,
+            to,
+            roomId
+          }
+        }}))
       })
-      this.setState({reservations})
     })
   }
 
-  changeDate = direction => {
-    this.setState( ({currentDate}) => ({
-      currentDate: !direction ? moment() : currentDate.add(direction, 'month')
-    }))
+
+  UNSAFE_componentWillReceiveProps = ({match:{params}}) => {
+    this.updateURL(params)
   }
 
-  handleDayClick = day => {
-    const {date, dayReservations} = day
-    const reservations = Object.assign({}, this.state.reservations)
-    Object.keys(reservations).forEach( key => {
-      !dayReservations.includes(key) && delete reservations[key]
-    })
-    
-    this.setState({
-      isDayBig: true,
-      date,
-      bigDayreservations: reservations
-    })
+  updateURL = ({
+    year, month
+  }) => {
+    this.setState({currentDate: moment().year(year).month(month-1)})
   }
 
-  closeBigDay = () => {
-    this.setState({isDayBig: false})
-  }
 
-  // FIXME: Clicking on hamburger menu triggers closeBigDay()
-  componentWillReceiveProps ({appBarRightAction}) {
-    if (appBarRightAction === "calendar") {
-      this.state.isDayBig ? this.closeBigDay() : this.changeDate(0)
-    }
-  }
-
-  render() {    
-    const reservations = {}
-    const {isDayBig, date, currentDate} = this.state
-    Object.entries(this.state.reservations || {}).forEach(reservation => {
-      const [key,value] = reservation
-      const {roomId, from, to} = value.metadata
-      reservations[key] = {roomId, from, to}
-    })
+  render() {
+    const {history} = this.props
+    const {
+      currentDate, reservations
+    } = this.state
 
 
     return (
       <div id="calendar-wrapper">
-        {isDayBig ?
-          <DayBig
-            closeBigDay={this.closeBigDay}
-            reservations={this.state.bigDayreservations}
-            {...{date}}
-          /> :
-          <Month 
-            handleDayClick={this.handleDayClick}
-            changeDate={this.changeDate}
-            {...{reservations, currentDate}}
-          />
-        }
+        <Route
+          component={
+            () => (
+              <Month
+                {...{
+                  history,
+                  reservations,
+                  currentDate
+                }}
+              />
+            )
+          }
+          exact
+          path={toRoute(routes.CALENDAR, ":year", ":month")}
+        />
+        <Route
+          path={toRoute(routes.CALENDAR, ":year", ":month", ":day")}
+          render={({
+            match, history
+          }) =>
+            <DayBig
+              {...{
+                match,
+                history
+              }}
+            />
+          }
+        />
       </div>
     )
   }
