@@ -1,89 +1,113 @@
 import React, {Component} from 'react'
+import moment from 'moment'
 
+import {ROOMS_DB} from '../../../utils/firebase'
+import {
+  TextField,
+  Grid,
+  Button,
+  Typography,
+  CardActions,
+  Card,
+  CardContent
+} from '@material-ui/core'
 
-import firebase from 'firebase'
-
-import Card from 'material-ui/Card'
-import Toggle from 'material-ui/Toggle'
-import Dialog from 'material-ui/Dialog'
-import RaisedButton from 'material-ui/RaisedButton'
+import Lock from '@material-ui/icons/LockRounded'
+import Unlock from '@material-ui/icons/LockOpenRounded'
 
 export default class Availability extends Component {
-  
+
   state = {
-    isAvailable: false,
-    isDialogOpen: false
+    unavailable: null,
+    unavailableTemp: null
   }
 
   componentDidMount() {
-    const roomAvailabilityRef = firebase.database().ref(`rooms/${this.props.roomId-1}/available`)
-    roomAvailabilityRef.on("value", snap => {
-      this.setState({
-        isAvailable: snap.val()
-      })
-    })
-  }
-  
-  
-  handleDialogOpen = () => {
-    this.setState({isDialogOpen: true})
+    ROOMS_DB.child(`${this.props.roomId-1}/unavailable`)
+      .on("value", snap => this.setState({unavailable: snap.val()}))
   }
 
-  handleDialogClose = () => this.setState({isDialogOpen: false})
+  handleChangeInput = ({target: {value}}) => this.setState({unavailableTemp: value})
 
-
-  handleAvailability = () => {
-    this.setState(({isAvailable}) => ({
-      isAvailable: !isAvailable
-    }), () => {
-      firebase.database().ref(`rooms/${this.props.roomId-1}/available`)
-      .set(this.state.isAvailable).then(() => this.handleDialogClose())
-    })
-
+  handleCancelBlock = () => {
+    this.props.openDialog({
+      content: "A szoba újra foglalható lesz.",
+      submitLabel: "Feloldás"
+    },
+    () => ROOMS_DB
+      .child(`${this.props.roomId-1}/unavailable`)
+      .remove(),
+    "Blokkolás feloldva. A szoba újra foglalható."
+    )
   }
 
-  
+  handleSubmitDate = () => {
+    this.props.openDialog({
+      submitLabel: "Blokkolás",
+      content: "A szoba foglalása blokkolva lesz az adott dátumig."
+    },
+    () => ROOMS_DB
+      .child(`${this.props.roomId-1}/unavailable`)
+      .set(this.state.unavailableTemp),
+    "A szoba blokkolva lett a megadott dátumig."
+    )
+  }
+
   render() {
-    const {roomId} = this.props
-    const {isAvailable, isDialogOpen} = this.state
+    const {unavailable, unavailableTemp} = this.state
+    const isUnavailable = moment(unavailable || undefined).startOf("day").isAfter(moment().startOf("day"))
     return (
-      <Card className="room-edit-block">
-        <div
-          style={{
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "center",
-            padding: "1em"
-          }}
-        >
-          <p>Szoba {roomId} jelenleg</p>
-          <Dialog 
-            title="Foglalás"
-            modal
-            open={isDialogOpen}
-            actions={[
-              <RaisedButton
-                label="Mégse"
-                onClick={this.handleDialogClose}
-              />,
-              <RaisedButton
-                style={{marginLeft: 12}}
-                secondary
-                label="Igen"
-                onClick={this.handleAvailability}
+      <Card>
+        <CardContent>
+          <Grid
+            alignItems="center"
+            container
+            justify="space-between"
+          >
+            <Typography variant="body2">
+              {isUnavailable ?
+                "Blokkolva eddig" :
+                "Nincs blokkolva"
+              }
+            </Typography>
+            {isUnavailable ?
+              <Typography>{moment(unavailable).format("YYYY. MMMM DD.")}</Typography> :
+              <TextField
+                onChange={this.handleChangeInput}
+                type="date"
+                value={moment(unavailableTemp || undefined).format("YYYY-MM-DD")}
               />
-            ]}
-          > Figyelem! A szobafoglalás ezzel <span style={{color: "red", fontWeight: "bold"}}>{isAvailable && "nem"} elérhetővé</span> válik. Biztos folytatja?
-          </Dialog>
-          <Toggle
-            label={<div>{!isAvailable && <span style={{fontWeight: "bold"}}>nem</span>} <span style={{fontWeight: isAvailable && "bold"}}>elérhető</span></div>}
-            onClick={() => this.handleDialogOpen()}
-            style={{width: "auto"}}
-            toggled={isAvailable}
-           
-          />
-        </div>
+            }
+          </Grid>
+        </CardContent>
+        <CardActions>
+          <Grid
+            container
+            justify="flex-end"
+          >
+            {isUnavailable ?
+              <Button
+                color="primary"
+                onClick={this.handleCancelBlock}
+                size="small"
+                variant="outlined"
+              >
+                <Unlock/>
+                Feloldás
+              </Button> :
+              <Button
+                color="secondary"
+                onClick={this.handleSubmitDate}
+                size="small"
+                variant="contained"
+              >
+                <Lock/>
+                Blokkol
+              </Button>
+            }
+          </Grid>
+        </CardActions>
       </Card>
-  )
+    )
   }
 }
