@@ -1,4 +1,8 @@
-import firebase from 'firebase'
+import firebase from "firebase/app"
+import "firebase/database"
+import "firebase/firestore"
+import "firebase/auth"
+import "firebase/storage"
 
 firebase.initializeApp({
   apiKey: "AIzaSyB4-Y2_RCdrOouJJxUJkBBXGyj4hNdjDs0",
@@ -8,3 +12,29 @@ firebase.initializeApp({
   storageBucket: "bibic-vendeghazak-api.appspot.com",
   messagingSenderId: "586582307718"
 })
+
+export const FS = firebase.firestore()
+FS.settings({timestampsInSnapshots: true})
+
+export const DB = firebase.database()
+export const FileStore = firebase.storage()
+export const AUTH = firebase.auth()
+
+
+export const RESERVATIONS_FS = FS.collection("reservations")
+export const GALLERIES_DB = DB.ref("galleries")
+export const FEEDBACKS_DB = DB.ref("feedbacks")
+export const ADMINS = DB.ref("admins")
+export const ROOMS_DB = DB.ref("rooms")
+export const FOODS_DB = DB.ref("foods")
+export const ROOM_SERVICES_DB = DB.ref("roomServices")
+export const SERVICES_DB = DB.ref("services")
+export const SERVER_MESSAGE_DB = DB.ref("serverMessage")
+export const RESERVATION_DATES_DB = DB.ref("reservationDates")
+export const PARAGRAPHS_DB = DB.ref("paragraphs")
+
+export const TIMESTAMP = firebase.firestore.FieldValue.serverTimestamp()
+
+export const getAdminName = uid => ADMINS
+  .child(`${uid}/name`)
+  .once("value")
